+++ packages/utilities/date-utils/src/index.ts
@@ -8,7 +8,7 @@ export * from "./format-visible-range"
 export * from "./get-decade-range"
 export * from "./get-month-dates"
 export * from "./get-month-names"
-export * from "./get-week-dates"
+export * from "./get-week-days"
 export * from "./get-year-range"
 export * from "./mutation"
 export * from "./pagination"
